Add initial comment thing implementation
Fix broken functions, so that things-bounds works with the aggregated
comment thing and initial tests (to be added).

The comment thing and a string thing (to be created) will be used in a
pair definer helper in order to ensure that any pair delimiters that
appear in a comment or string must be matched in that same comment or
string (or they will ignored/not be considered part of a pair).